Fix AutoRefreshbutton tests
